backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart

sentence-transformers
faiss-cpu
PyPDF2
python-docx

python-dotenv

langchain
langchain-core
langchain-community

generative-ai-hub-sdk    # <-- correct working package

requests
